REF: avoid circular import of get_adjustment
